requirements.txt
tensorflow>=2.13.0
torch>=1.9.0
torchvision>=0.10.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
Pillow>=8.3.1
numpy>=1.21.0
pandas>=2.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
python-dotenv>=1.0.0
pymongo>=4.4.0
motor>=3.3.0
requests>=2.31.0
opencv-python>=4.8.0
albumentations>=1.3.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic[email]>=2.0.0
bson>=0.5.10
tqdm>=4.61.2 